自分の日記データを統計カードとして表示可能に
